增加获取任务数功能

+++ src/main/java/com/example/todolist/pojo/Task.java
@@ -22,7 +22,7 @@ public class Task {
     private Date beginTime;
     private Date endTime;
     private Integer state;
-    private Double finishRate;
+    private Integer finishRate;
 
     @Override
     public String toString() {
